pyproject.toml
[project]
name = "practice-python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pyhumps>=3.8.0",
    "pymongo>=4.14.0",
]
